에러 핸들링 코드 일괄 적용

+++ src/main/java/kr/ac/deu/computer_engineering/Absenteeism/Management/route/company/CompanyController.java
@@ -4,15 +4,21 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.Company.Company;
 import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.Company.dto.CompanyDto;
-import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.Company.dto.CompanyDto;
 import kr.ac.deu.computer_engineering.Absenteeism.Management.service.company.CompanyService;
+import kr.ac.deu.computer_engineering.Absenteeism.Management.utils.RoleValidate;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+import javax.validation.Valid;
 import java.util.List;
 
+@Slf4j
 @RestController
 @RequiredArgsConstructor
 @RequestMapping(value = "/api/company")
@@ -24,7 +30,9 @@ public class CompanyController {
             summary = "회사 목록 조회",
             description = "회사 목록 조회")
     @GetMapping("")
-    public ResponseEntity<?> getItemList(@RequestParam(required = false) String q) {
+    public ResponseEntity<?> getItemList(
+            HttpServletRequest request,
+            @RequestParam(required = false) String q) {
         List<Company> result = companyService.getList(q);
         return new ResponseEntity<>(result, HttpStatus.OK);
     }
@@ -34,9 +42,16 @@ public class CompanyController {
             summary = "회사 Id로 상세 조회",
             description = "회사 상세 조회")
     @GetMapping("/{companyId}")
-    public ResponseEntity<?> getItemById(@PathVariable Long companyId) throws Exception {
-        Company comp = companyService.getCompanyById(companyId);
-        return new ResponseEntity<>(comp, HttpStatus.OK);
+    public ResponseEntity<?> getItemById(
+            HttpServletRequest request,
+            @PathVariable Long companyId) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session)) {
+            Company comp = companyService.getCompanyById(companyId);
+            return new ResponseEntity<>(comp, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @Tag(name = "회사")
@@ -45,10 +60,17 @@ public class CompanyController {
             description = "회사 정보 생성")
     @PostMapping("")
     public ResponseEntity<?> createItem(
-            @RequestBody CompanyDto dto
-    ) throws Exception {
-        companyService.createCompany(dto);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+            HttpServletRequest request,
+            @Valid @RequestBody CompanyDto dto,
+            BindingResult bindingResult
+    ) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session)) {
+            companyService.createCompany(dto);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @Tag(name = "회사")
@@ -57,10 +79,17 @@ public class CompanyController {
             description = "회사 정보 수정")
     @PutMapping("/{companyId}")
     public ResponseEntity<?> updateItemById(
-            @RequestBody CompanyDto dto,
-            @PathVariable Long companyId) throws Exception {
-        companyService.updateCompany(companyId, dto);
-        return new ResponseEntity<>(HttpStatus.OK);
+            HttpServletRequest request,
+            @PathVariable Long companyId,
+            @Valid @RequestBody CompanyDto dto,
+            BindingResult bindingResult) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session)) {
+            companyService.updateCompany(companyId, dto);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @Tag(name = "회사")
@@ -68,8 +97,15 @@ public class CompanyController {
             summary = "회사 Id로 직급 정보 삭제",
             description = "회사 정보 삭제")
     @DeleteMapping("/{companyId}")
-    public ResponseEntity<?> deleteItemById(@PathVariable Long companyId) throws Exception {
-        companyService.deleteCompany(companyId);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<?> deleteItemById(
+            HttpServletRequest request,
+            @PathVariable Long companyId) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session)) {
+            companyService.deleteCompany(companyId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 }
